nrfs: put mtime, add gen directly in item data
mtime is frequently updated. To speed things up it is put directly in
the item to avoid frequent attribute storage reallocation.

A generation field has been added. It will be used for efficient update
tracking for backup tools etc.

+++ fuse/src/fs/mod.rs
@@ -192,13 +192,13 @@ impl Fs {
 	}
 }
 
-fn mtime_now() -> i128 {
+fn mtime_now() -> i64 {
 	mtime_sys(SystemTime::now())
 }
 
-fn mtime_sys(t: SystemTime) -> i128 {
+fn mtime_sys(t: SystemTime) -> i64 {
 	t.duration_since(UNIX_EPOCH).map_or_else(
-		|t| -t.duration().as_micros().try_into().unwrap_or(i128::MAX),
-		|t| t.as_micros().try_into().unwrap_or(i128::MAX),
+		|t| -t.duration().as_micros().try_into().unwrap_or(i64::MAX),
+		|t| t.as_micros().try_into().unwrap_or(i64::MAX),
 	)
 }
